requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
newsapi-python==0.2.6
pymongo==3.12.1
python-dotenv==0.19.1
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
